+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adds support for 204 no content in generation and CSharp/Java/Go/TypeScript request adapters. #1410
+
 ### Changed
 
 - Fixed a bug where request builder classes for collections endpoints would have a wrong name. #1052